+++ packages/box-tracker/lib/const.ts
@@ -1,1 +1,1 @@
-export const MAX_BOX_HEIGHT = 10;
+export const MAX_BOX_LENGTH = 10;