+++ packages/box-tracker/lib/interfaces.ts
@@ -1,34 +1,17 @@
-import {
-  AdditionalRegisters,
-  Asset,
-  Transaction,
-} from '@rosen-bridge/scanner-interfaces';
+import { OutputBox, Transaction } from '@rosen-bridge/scanner-interfaces';
 
 export type Token = {
   tokenId: string;
   amount: bigint;
 };
 
-export type ErgoBox = {
-  boxId: string;
-  value: bigint;
-  ergoTree: string;
-  creationHeight: number;
-  assets: Array<Asset>;
-  additionalRegisters: AdditionalRegisters;
-  transactionId: string;
-  index: number;
-};
+export interface ErgoBox extends OutputBox {
+  blockId: string;
+}
 
 export interface MempoolTrackResult {
-  boxes: ErgoBox[];
+  boxes: OutputBox[];
   spentBoxIds: string[];
 }
 
 export type TxDeserializer = (serializedTx: string) => Transaction;
-
-export type BoxWithHeight = {
-  box: ErgoBox;
-  inclusionHeight: number;
-  hash: string;
-};