+++ packages/raw-data-provider/lib/migrations/sqlite/index.ts
@@ -1,0 +1,1 @@
+export { Migration1762178166781 } from './1762178166781-migration';