+++ packages/abstract-scanner/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @rosen-bridge/abstract-scanner
 
+## 0.2.3
+
+### Patch Changes
+
+- Fix Web-socket scanner mutex release after max retrials
+- Update dependencies
+  - @rosen-bridge/rosen-extractor@11.1.1
+
+## 0.2.2
+
+### Patch Changes
+
+- Fixed type definitions in entities based migrations
+
 ## 0.2.1
 
 ### Patch Changes