+++ packages/observation-extractors/cardano-observation-extractor/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @rosen-bridge/cardano-observation-extractor
 
+## 1.1.1
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-observation-extractor@0.2.3
+  - @rosen-bridge/rosen-extractor@11.1.1
+
+## 1.1.0
+
+### Minor Changes
+
+- Add `storeRawData` option to store observations raw-data into the database
+
+### Patch Changes
+
+- Update Dependencies
+  - @rosen-bridge/abstract-observation-extractor@0.2.2
+  - @rosen-bridge/rosen-extractor@11.1.0
+
+## 1.0.0
+
+### Major Changes
+
+- Change type of metadata for Blockfrost
+
+### Patch Changes
+
+- Update Dependencies
+  - @rosen-bridge/abstract-observation-extractor@0.2.1
+  - @rosen-bridge/rosen-extractor@11.0.0
+
 ## 0.2.0
 
 ### Minor Changes