+++ packages/scanners/bitcoin-scanner/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @rosen-bridge/bitcoin-scanner
 
+## 0.2.3
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-scanner@0.2.3
+
+## 0.2.2
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-scanner@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes