+++ packages/scanners/evm-scanner/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @rosen-bridge/evm-scanner
 
+## 0.1.4
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-scanner@0.2.3
+
+## 0.1.3
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-scanner@0.2.2
+
 ## 0.1.2
 
 ### Patch Changes